REFACTOR: BB-1521 Move and reorganize PivotTable code
