+++ src/components/core/pivotTable/agGridConst.ts
@@ -1,0 +1,13 @@
+// (C) 2007-2019 GoodData Corporation
+
+export const ROW_ATTRIBUTE_COLUMN = "ROW_ATTRIBUTE_COLUMN";
+export const COLUMN_ATTRIBUTE_COLUMN = "COLUMN_ATTRIBUTE_COLUMN";
+export const MEASURE_COLUMN = "MEASURE_COLUMN";
+export const FIELD_SEPARATOR = "-";
+export const FIELD_SEPARATOR_PLACEHOLDER = "DASH";
+export const FIELD_TYPE_MEASURE = "m";
+export const FIELD_TYPE_ATTRIBUTE = "a";
+export const ID_SEPARATOR = "_";
+export const ID_SEPARATOR_PLACEHOLDER = "UNDERSCORE";
+export const DOT_PLACEHOLDER = "DOT";
+export const ROW_TOTAL = "rowTotal";
