+++ src/components/core/pivotTable/agGridDataSource.ts
@@ -2,16 +2,31 @@
 import { AFM, Execution } from "@gooddata/typings";
 import isEqual = require("lodash/isEqual");
 import { IDatasource, IGetRowsParams, GridApi } from "ag-grid";
+import { getMappingHeaderName } from "../../../helpers/mappingHeader";
 
 import InjectedIntl = ReactIntl.InjectedIntl;
 
-import { executionToAGGridAdapter, ROW_ATTRIBUTE_COLUMN } from "../../../helpers/agGrid";
-import { IGridHeader, IGridTotalsRow } from "../../../interfaces/AGGrid";
+import { getTreeLeaves, getSubtotalStyles } from "./agGridUtils";
+import { ROW_ATTRIBUTE_COLUMN } from "./agGridConst";
+import {
+    getMeasureSortItemFieldAndDirection,
+    getSortsFromModel,
+    getAttributeSortItemFieldAndDirection,
+    assignSorting,
+} from "./agGridSorting";
+import { IAgGridPage, IGridAdapterOptions, IGridHeader, IGridTotalsRow } from "./agGridTypes";
 
 import ApiWrapper from "./agGridApiWrapper";
 import { IGetPage } from "../base/VisualizationLoadingHOC";
 import { IGroupingProvider } from "../pivotTable/GroupingProvider";
-import { getSortsFromModel } from "../PivotTable";
+import {
+    assortDimensionHeaders,
+    getColumnHeaders,
+    getFields,
+    getMinimalRowData,
+    getRowHeaders,
+} from "./agGridHeaders";
+import { getRow, getRowTotals } from "./agGridData";
 
 export const areTotalsChanged = (gridApi: GridApi, newTotals: IGridTotalsRow[]) => {
     const currentTotalsCount = gridApi.getPinnedBottomRowCount();
@@ -123,6 +138,122 @@ export const getDataSourceRowsGetter = (
 
             return execution;
         });
+    };
+};
+
+export const executionToAGGridAdapter = (
+    executionResponses: Execution.IExecutionResponses,
+    resultSpec: AFM.IResultSpec = {},
+    intl: InjectedIntl,
+    options: IGridAdapterOptions = {},
+): IAgGridPage => {
+    const { makeRowGroups = false, addLoadingRenderer = null, columnDefOptions } = options;
+
+    const {
+        executionResponse: { dimensions },
+        executionResult: { data, headerItems, totals },
+    } = executionResponses;
+
+    const columnAttributeHeaderCount = dimensions[1].headers.filter(
+        (header: Execution.IHeader) => !!(header as Execution.IAttributeHeader).attributeHeader,
+    ).length;
+
+    const columnHeaders: IGridHeader[] = getColumnHeaders(
+        headerItems[1],
+        dimensions[1].headers,
+        columnDefOptions,
+    );
+    const groupColumnHeaders: IGridHeader[] =
+        columnAttributeHeaderCount > 0
+            ? [
+                  {
+                      headerName: dimensions[1].headers
+                          .filter(header => Execution.isAttributeHeader(header))
+                          .map((header: Execution.IAttributeHeader) => {
+                              return getMappingHeaderName(header);
+                          })
+                          .filter((item: string) => item !== null)
+                          .join(" › "),
+                      field: "columnGroupLabel",
+                      children: columnHeaders,
+                      drillItems: [],
+                  },
+              ]
+            : columnHeaders;
+
+    const rowHeaders: IGridHeader[] =
+        // There are supposed to be only attribute headers on the first dimension
+        getRowHeaders(dimensions[0].headers as Execution.IAttributeHeader[], columnDefOptions, makeRowGroups);
+
+    // build sortingMap from resultSpec.sorts
+    const sorting = resultSpec.sorts || [];
+    const sortingMap = {};
+    const { attributeHeaders, measureHeaderItems } = assortDimensionHeaders(dimensions);
+    sorting.forEach(sortItem => {
+        if (AFM.isAttributeSortItem(sortItem)) {
+            const [field, direction] = getAttributeSortItemFieldAndDirection(sortItem, attributeHeaders);
+            sortingMap[field] = direction;
+        }
+        if (AFM.isMeasureSortItem(sortItem)) {
+            const [field, direction] = getMeasureSortItemFieldAndDirection(sortItem, measureHeaderItems);
+            sortingMap[field] = direction;
+        }
+    });
+    // assign sorting and indexes
+    const columnDefs: IGridHeader[] = [...rowHeaders, ...groupColumnHeaders].map((column, index) => {
+        if (column.children) {
+            getTreeLeaves(column).forEach((leafColumn, leafColumnIndex) => {
+                leafColumn.index = index + leafColumnIndex;
+                assignSorting(leafColumn, sortingMap);
+            });
+        }
+        column.index = index;
+        assignSorting(column, sortingMap);
+        return column;
+    });
+
+    // Add loading indicator to the first column
+    if (addLoadingRenderer) {
+        const leafColumnDefs = getTreeLeaves(columnDefs);
+        if (leafColumnDefs.length > 0) {
+            leafColumnDefs[0].cellRenderer = addLoadingRenderer;
+        }
+    }
+
+    const columnFields: string[] = getFields(headerItems[1]);
+    const rowFields: string[] = rowHeaders.map(header => header.field);
+    // PivotTable execution should always return a two-dimensional array (Execution.DataValue[][])
+    const minimalRowData: Execution.DataValue[][] = getMinimalRowData(
+        data as Execution.DataValue[][],
+        headerItems[0],
+    );
+
+    const rowData = minimalRowData.map((dataRow: Execution.DataValue[], dataRowIndex: number) =>
+        getRow(
+            dataRow,
+            dataRowIndex,
+            columnFields,
+            rowHeaders,
+            headerItems[0],
+            getSubtotalStyles(resultSpec.dimensions ? resultSpec.dimensions[0] : null),
+            intl,
+        ),
+    );
+
+    const columnKeys = [...rowFields, ...columnFields];
+    const rowTotals = getRowTotals(
+        totals,
+        columnKeys,
+        dimensions[0].headers,
+        resultSpec,
+        measureHeaderItems.map(mhi => mhi.measureHeaderItem.localIdentifier),
+        intl,
+    );
+
+    return {
+        columnDefs,
+        rowData,
+        rowTotals,
     };
 };
