+++ src/components/core/pivotTable/tests/agGridDataSource.spec.ts
@@ -1,10 +1,11 @@
 // (C) 2019 GoodData Corporation
-import { IGridTotalsRow } from "../../../../interfaces/AGGrid";
-import { areTotalsChanged, getAGGridDataSource } from "../agGridDataSource";
-import { pivotTableWithColumnAndRowAttributes } from "../../../../../stories/test_data/fixtures";
+import * as fixtures from "../../../../../stories/test_data/fixtures";
+import { IGridTotalsRow } from "../agGridTypes";
+import { areTotalsChanged, getAGGridDataSource, executionToAGGridAdapter } from "../agGridDataSource";
 import { GroupingProviderFactory } from "../GroupingProvider";
 import { createIntlMock } from "../../../visualizations/utils/intlUtils";
 
+const pivotTableWithColumnAndRowAttributes = fixtures.pivotTableWithColumnAndRowAttributes;
 const intl = createIntlMock();
 
 describe("getGridDataSource", () => {
@@ -156,5 +157,32 @@ describe("getGridDataSource", () => {
                 expect(areTotalsChanged(gridApi, passedTotals)).toBe(expectedValue);
             },
         );
+    });
+});
+
+describe("executionToAGGridAdapter", () => {
+    it("should return grid data for executionResult", () => {
+        expect(
+            executionToAGGridAdapter(
+                {
+                    executionResponse: fixtures.pivotTableWithColumnAndRowAttributes.executionResponse,
+                    executionResult: fixtures.pivotTableWithColumnAndRowAttributes.executionResult,
+                },
+                {},
+                intl,
+            ),
+        ).toMatchSnapshot();
+    });
+    it("should return grid data for executionResult with rowGroups and loadingRenderer", () => {
+        expect(
+            executionToAGGridAdapter(
+                {
+                    executionResponse: fixtures.barChartWithStackByAndOnlyOneStack.executionResponse,
+                    executionResult: fixtures.barChartWithStackByAndOnlyOneStack.executionResult,
+                },
+                {},
+                intl,
+            ),
+        ).toMatchSnapshot();
     });
 });
