--- src/helpers/agGridFixtures.ts
@@ -1,1011 +1,0 @@
-// (C) 2007-2018 GoodData Corporation
-export const executionResult = {
-    result: {
-        executionResponse: {
-            dimensions: [
-                {
-                    headers: [
-                        {
-                            attributeHeader: {
-                                name: "Location State",
-                                localIdentifier: "locationState",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2211",
-                                identifier: "label.restaurantlocation.locationstate",
-                                formOf: {
-                                    name: "Location State",
-                                    uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210",
-                                    identifier: "attr.restaurantlocation.locationstate",
-                                },
-                            },
-                        },
-                        {
-                            attributeHeader: {
-                                name: "Location Name",
-                                localIdentifier: "locationName",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2205",
-                                identifier: "label.restaurantlocation.locationname",
-                                formOf: {
-                                    name: "Location Name",
-                                    uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204",
-                                    identifier: "attr.restaurantlocation.locationname",
-                                },
-                            },
-                        },
-                    ],
-                },
-                {
-                    headers: [
-                        {
-                            attributeHeader: {
-                                name: "Year (Date)",
-                                localIdentifier: "year",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2005",
-                                identifier: "date.aag81lMifn6q",
-                                formOf: {
-                                    name: "Year (Date)",
-                                    uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003",
-                                    identifier: "date.year",
-                                },
-                            },
-                        },
-                        {
-                            attributeHeader: {
-                                name: "US Short (Date)",
-                                localIdentifier: "quarter",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2131",
-                                identifier: "date.aci81lMifn6q",
-                                formOf: {
-                                    name: "Quarter/Year (Date)",
-                                    uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129",
-                                    identifier: "date.quarter",
-                                },
-                            },
-                        },
-                        {
-                            measureGroupHeader: {
-                                items: [
-                                    {
-                                        measureHeaderItem: {
-                                            name: "$ Total Sales",
-                                            format: "[>=0]$#,##0;[<0]-$#,##0",
-                                            localIdentifier: "totalSales",
-                                            uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2352",
-                                            identifier: "aa7ulGyKhIE5",
-                                        },
-                                    },
-                                    {
-                                        measureHeaderItem: {
-                                            name: "Franchise fees",
-                                            format: "[>=0]$#,##0;[<0]-$#,##0",
-                                            localIdentifier: "franchiseFees",
-                                            uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/6685",
-                                            identifier: "aaEGaXAEgB7U",
-                                        },
-                                    },
-                                ],
-                            },
-                        },
-                    ],
-                },
-            ],
-            links: {
-                // tslint:disable-next-line:max-line-length
-                executionResult:
-                    "/gdc/app/projects/xms7ga4tf3g3nzucd8380o2bev8oeknp/executionResults/6702535957456288768?q=eAGtVV9v0zAQ%2FyqRu4chZYrjtFsYDDSpFE1CSKPbA4ry4MbXEurYkeMwxtTvznlpAqTwQNI3%2B%2Byc%0A735%2FLk%2FEQKmN%2FcgLIJfkXtncShDEJ5mWdaEqcpk8EZFXpeSPC22KG4HXgo3IgkIE34vqYsOndh1t%0AIvWjzkQcxVSzFXyLNWxVGejV14BROsN8K17BOwkFKHv%2F6WZAkiiA5vPqbS6uMCO31uSr2sKgmiLM%0AIEwu5Vw%2FqKWF8rpNN1%2F8d3VhFGK6CoHET8utW9dFwc0jbj8DN79Xe%2BcQ3se90zm38AKPn3E%2FjGrL%0ApeMgSf0kTXf%2BWC6aQkdyEbKXx%2BQC0yEAx%2BNiyhz%2Bf%2BXitubGgvG4Ep7jxTu9DVGgIXUUdHpqGdrf%0ADpqbfab%2BcdpnjBSAMs3eG12XJPX328ZXazQUd5JJ3lzR9GTiTyb0VfKapmfN2vXR6Qg3rUhOvDv3%0AjLfkEiqMN08MsEE0Y8SJamghC8NV9iWvwFvDqErOz%2BMZ2SE8Rj802Byw8UFn3OZaeUuLnulDM25E%0AsdD5d6QtWEj7tmgJO6i9624AafiO00JfaGNHAzvGmGZ02sdgVKt0iq3KXG2vh0%2FnKGYuyx9O6opq%0Avd4x9Pwn%2FGW1g3gP9nT3E1A7Yf8%3D%0A&c=d3ae6c3186fa60b93e8cca68cffe2595&offset=0%2C0&limit=1000%2C1000&dimensions=2&totals=0%2C0",
-            },
-        },
-        executionResult: {
-            data: [
-                [
-                    "1374718.75",
-                    "156851.09375",
-                    "1372349.8",
-                    "156649.733",
-                    "1341183.75",
-                    "154000.61875",
-                    "1363659.55",
-                    "155911.06175",
-                    "1374718.75",
-                    "156851.09375",
-                    "1372349.8",
-                    "156649.733",
-                    "1341183.75",
-                    "154000.61875",
-                    "1363659.55",
-                    "155911.06175",
-                    "1374718.75",
-                    "156851.09375",
-                    "1372349.8",
-                    "156649.733",
-                    "1341183.75",
-                    "154000.61875",
-                    "1084960.15",
-                    "132221.61275",
-                ],
-                [
-                    "798418.27",
-                    null,
-                    "804066.97",
-                    null,
-                    "786092.26",
-                    null,
-                    "797158.88",
-                    null,
-                    "798418.27",
-                    null,
-                    "804066.97",
-                    null,
-                    "786092.26",
-                    null,
-                    "797158.88",
-                    null,
-                    "798418.27",
-                    null,
-                    "804066.97",
-                    null,
-                    "786092.26",
-                    null,
-                    "632411.13",
-                    null,
-                ],
-                [
-                    "709310.82",
-                    null,
-                    "711655.42",
-                    null,
-                    "690235.37",
-                    null,
-                    "699431.6",
-                    null,
-                    "709310.82",
-                    null,
-                    "711655.42",
-                    null,
-                    "690235.37",
-                    null,
-                    "699431.6",
-                    null,
-                    "709310.82",
-                    null,
-                    "711655.42",
-                    null,
-                    "690235.37",
-                    null,
-                    "555203.41",
-                    null,
-                ],
-                [
-                    "719950.54",
-                    "101195.7959",
-                    "721447.09",
-                    "101323.00265",
-                    "702665.02",
-                    "99726.5267",
-                    "714055.21",
-                    "100694.69285",
-                    "719950.54",
-                    "101195.7959",
-                    "721447.09",
-                    "101323.00265",
-                    "702665.02",
-                    "99726.5267",
-                    "714055.21",
-                    "100694.69285",
-                    "719950.54",
-                    "101195.7959",
-                    "721447.09",
-                    "101323.00265",
-                    "702665.02",
-                    "99726.5267",
-                    "568202.31",
-                    "88297.19635",
-                ],
-                [
-                    "665403.4",
-                    "96559.289",
-                    "664961.05",
-                    "96521.68925",
-                    "647681.4",
-                    "95052.919",
-                    "661683.8",
-                    "96243.123",
-                    "665403.4",
-                    "96559.289",
-                    "664961.05",
-                    "96521.68925",
-                    "647681.4",
-                    "95052.919",
-                    "661683.8",
-                    "96243.123",
-                    "665403.4",
-                    "96559.289",
-                    "664961.05",
-                    "96521.68925",
-                    "647681.4",
-                    "95052.919",
-                    "526652.15",
-                    "84765.43275",
-                ],
-                [
-                    "521114.63",
-                    "84294.74355",
-                    "518945.99",
-                    "84110.40915",
-                    "505723.75",
-                    "82986.51875",
-                    "511277.94",
-                    "83458.6249",
-                    "521114.63",
-                    "84294.74355",
-                    "518945.99",
-                    "84110.40915",
-                    "505723.75",
-                    "82986.51875",
-                    "511277.94",
-                    "83458.6249",
-                    "521114.63",
-                    "84294.74355",
-                    "518945.99",
-                    "84110.40915",
-                    "505723.75",
-                    "82986.51875",
-                    "407440.14",
-                    "74632.4119",
-                ],
-                [
-                    "756386.46",
-                    null,
-                    "760153.5",
-                    null,
-                    "735571.38",
-                    null,
-                    "751096.47",
-                    null,
-                    "756386.46",
-                    null,
-                    "760153.5",
-                    null,
-                    "735571.38",
-                    null,
-                    "751096.47",
-                    null,
-                    "756386.46",
-                    null,
-                    "760153.5",
-                    null,
-                    "735571.38",
-                    null,
-                    "598467.19",
-                    null,
-                ],
-                [
-                    "384607.25",
-                    "72691.61625",
-                    "382295.25",
-                    "72495.09625",
-                    "374582",
-                    "71839.47",
-                    "376215.85",
-                    "71978.34725",
-                    "384607.25",
-                    "72691.61625",
-                    "382295.25",
-                    "72495.09625",
-                    "374582",
-                    "71839.47",
-                    "376215.85",
-                    "71978.34725",
-                    "384607.25",
-                    "72691.61625",
-                    "382295.25",
-                    "72495.09625",
-                    "374582",
-                    "71839.47",
-                    "299928.4",
-                    "65493.914",
-                ],
-                [
-                    "708583.02",
-                    null,
-                    "705397.76",
-                    null,
-                    "689448.33",
-                    null,
-                    "699509.63",
-                    null,
-                    "708583.02",
-                    null,
-                    "705397.76",
-                    null,
-                    "689448.33",
-                    null,
-                    "699509.63",
-                    null,
-                    "708583.02",
-                    null,
-                    "705397.76",
-                    null,
-                    "689448.33",
-                    null,
-                    "556655.82",
-                    null,
-                ],
-                [
-                    "346719.56",
-                    null,
-                    "346551.26",
-                    null,
-                    "339011.88",
-                    null,
-                    "342288.82",
-                    null,
-                    "346719.56",
-                    null,
-                    "346551.26",
-                    null,
-                    "339011.88",
-                    null,
-                    "342288.82",
-                    null,
-                    "346719.56",
-                    null,
-                    "346551.26",
-                    null,
-                    "339011.88",
-                    null,
-                    "272910.32",
-                    null,
-                ],
-                [
-                    "395049.65",
-                    null,
-                    "391794.49",
-                    null,
-                    "382130.73",
-                    null,
-                    "387345.64",
-                    null,
-                    "395049.65",
-                    null,
-                    "391794.49",
-                    null,
-                    "382130.73",
-                    null,
-                    "387345.64",
-                    null,
-                    "395049.65",
-                    null,
-                    "391794.49",
-                    null,
-                    "382130.73",
-                    null,
-                    "307503.35",
-                    null,
-                ],
-                [
-                    "532091.15",
-                    "85227.74775",
-                    "538608.4",
-                    "85781.714",
-                    "529155.25",
-                    "84978.19625",
-                    "527335.8",
-                    "84823.543",
-                    "532091.15",
-                    "85227.74775",
-                    "538608.4",
-                    "85781.714",
-                    "529155.25",
-                    "84978.19625",
-                    "527335.8",
-                    "84823.543",
-                    "532091.15",
-                    "85227.74775",
-                    "538608.4",
-                    "85781.714",
-                    "529155.25",
-                    "84978.19625",
-                    "421939.75",
-                    "75864.87875",
-                ],
-            ],
-            paging: {
-                count: [12, 24],
-                offset: [0, 0],
-                total: [12, 24],
-            },
-            headerItems: [
-                [
-                    [
-                        {
-                            attributeHeaderItem: {
-                                name: "Alabama",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=6340109",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "California",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=6340116",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "California",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=6340116",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "California",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=6340116",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "California",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=6340116",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "California",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=6340116",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Florida",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=6340105",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Florida",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=6340105",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "New York",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=6340112",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "New York",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=6340112",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Texas",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=4436534",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Texas",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2210/elements?id=4436534",
-                            },
-                        },
-                    ],
-                    [
-                        {
-                            attributeHeaderItem: {
-                                name: "Montgomery",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340107",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Daly City",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340114",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Hayward",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340119",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Highland Village",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340121",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "San Jose - Blossom Hill",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340125",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "San Jose - Saratoga",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340127",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Aventura",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340103",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Deerfield Beach",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340117",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Manhattan - Harlem",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340113",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Times Square",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340128",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Dallas - Lemmon Avenue",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=5703467",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Irving",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2204/elements?id=6340132",
-                            },
-                        },
-                    ],
-                ],
-                [
-                    [
-                        {
-                            attributeHeaderItem: {
-                                name: "2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2015",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2015",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2015",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2015",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2015",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2015",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2015",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2015",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2016",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2016",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2016",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2016",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2016",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2016",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2016",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2016",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2017",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2017",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2017",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2017",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2017",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2017",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2017",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2003/elements?id=2017",
-                            },
-                        },
-                    ],
-                    [
-                        {
-                            attributeHeaderItem: {
-                                name: "Q1/2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8061",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q1/2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8061",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q2/2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8062",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q2/2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8062",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q3/2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8063",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q3/2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8063",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q4/2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8064",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q4/2015",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8064",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q1/2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8065",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q1/2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8065",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q2/2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8066",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q2/2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8066",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q3/2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8067",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q3/2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8067",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q4/2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8068",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q4/2016",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8068",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q1/2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8069",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q1/2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8069",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q2/2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8070",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q2/2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8070",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q3/2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8071",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q3/2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8071",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q4/2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8072",
-                            },
-                        },
-                        {
-                            attributeHeaderItem: {
-                                name: "Q4/2017",
-                                uri: "/gdc/md/xms7ga4tf3g3nzucd8380o2bev8oeknp/obj/2129/elements?id=8072",
-                            },
-                        },
-                    ],
-                    [
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "$ Total Sales",
-                                order: 0,
-                            },
-                        },
-                        {
-                            measureHeaderItem: {
-                                name: "Franchise fees",
-                                order: 1,
-                            },
-                        },
-                    ],
-                ],
-            ],
-        },
-    },
-};
